+++ .gitignore
@@ -1,0 +1,2 @@
+template_c/*
+dev/*
