+++ .htaccess
@@ -3,8 +3,8 @@
 
 # .htaccessでRewriteを有効化する
 RewriteEngine On
-RewriteCond %{HTTPS} off
-RewriteRule ^(.*)$ https://%{HTTP_HOST}%{REQUEST_URI} [R=301,L]
+#RewriteCond %{HTTPS} off
+#RewriteRule ^(.*)$ https://%{HTTP_HOST}%{REQUEST_URI} [R=301,L]
 ErrorDocument 400 /common/error/400/
 ErrorDocument 403 /common/error/403/
 ErrorDocument 404 /common/error/404/
