+++ .user.ini
@@ -1,2 +1,2 @@
-display_errors = Off
+display_errors = On
 error_reporting = E_ALL & ~E_NOTICE
