+++ public/Class/subdirectory/templates_c/518b91510401123a130defac21dc62a59b529470_0.file.index.tpl.php
@@ -1,0 +1,26 @@
+<?php
+/* Smarty version 3.1.32, created on 2018-09-11 20:00:42
+  from 'C:\Users\bokku\Documents\Project\bokkun.jp.project\public\Class\index.tpl' */
+
+/* @var Smarty_Internal_Template $_smarty_tpl */
+if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
+  'version' => '3.1.32',
+  'unifunc' => 'content_5b9802cae71f64_34440713',
+  'has_nocache_code' => false,
+  'file_dependency' => 
+  array (
+    '518b91510401123a130defac21dc62a59b529470' => 
+    array (
+      0 => 'C:\\Users\\bokku\\Documents\\Project\\bokkun.jp.project\\public\\Class\\index.tpl',
+      1 => 1534098063,
+      2 => 'file',
+    ),
+  ),
+  'includes' => 
+  array (
+  ),
+),false)) {
+function content_5b9802cae71f64_34440713 (Smarty_Internal_Template $_smarty_tpl) {
+?>ようこそ、<?php echo $_smarty_tpl->tpl_vars['name']->value;?>
+さん。<?php }
+}
